Add Login Page and User Access Control

+++ base/views.py
@@ -1,14 +1,49 @@
+from django.http import HttpResponse
 from django.shortcuts import render, redirect
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
 from django.db.models import Q
 from .models import Room, Topic
 from .form import RoomForm
 
 # Create your views here.
-rooms = [
-    {'id': 1, 'name': 'Room 1', 'description': 'Room 1 description', },
-    {'id': 2, 'name': 'Room 2', 'description': 'Room 2 description', },
-    {'id': 3, 'name': 'Room 3', 'description': 'Room 3 description', },
-]
+# rooms = [
+#     {'id': 1, 'name': 'Room 1', 'description': 'Room 1 description', },
+#     {'id': 2, 'name': 'Room 2', 'description': 'Room 2 description', },
+#     {'id': 3, 'name': 'Room 3', 'description': 'Room 3 description', },
+# ]
+
+
+def loginUser(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, "User does not exist")
+
+        user = authenticate(request, username=username, password=password)
+
+        print("User", user)
+        if user is not None:
+            login(request, user)
+            messages.success(request, "Logged in successfully")
+            return redirect('home')
+
+    return render(request, 'base/login_registration.html', {})
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
 
 
 def home(request):
@@ -38,6 +73,7 @@ def room(request, pk):
     return render(request, 'base/room.html', context=context)
 
 
+@login_required(login_url='login')
 def createRoom(request):
     room_form = RoomForm()
     if request.method == 'POST':
@@ -50,8 +86,13 @@ def createRoom(request):
     return render(request, 'base/room_form.html', context=context)
 
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user.username != room.host.username:
+        return HttpResponse('You are not allowed here.')
+
     # Here we given the instance to load in the screen
     form = RoomForm(instance=room)
 
@@ -65,8 +106,12 @@ def updateRoom(request, pk):
     return render(request, 'base/room_form.html', context=context)
 
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user.username != room.host.username:
+        return HttpResponse('You are not allowed here.')
 
     if request.method == 'POST':
         room.delete()
